refactor: Remove messages related functions from SharedUtils
