+++ app/src/main/java/com/infomaniak/mail/ui/main/thread/ThreadViewModel.kt
@@ -59,6 +59,7 @@ import com.infomaniak.mail.ui.main.thread.models.EmojiReactionAuthorUi
 import com.infomaniak.mail.ui.main.thread.models.EmojiReactionStateUi
 import com.infomaniak.mail.ui.main.thread.models.MessageUi
 import com.infomaniak.mail.ui.main.thread.models.MessageUi.UnsubscribeState
+import com.infomaniak.mail.useCases.MessagesActionsUseCase
 import com.infomaniak.mail.utils.AccountUtils
 import com.infomaniak.mail.utils.FeatureAvailability
 import com.infomaniak.mail.utils.FeatureAvailability.isSnoozeAvailable
@@ -123,6 +124,7 @@ class ThreadViewModel @Inject constructor(
     private val mailboxContentRealm: RealmDatabase.MailboxContent,
     private val mailboxController: MailboxController,
     private val messageController: MessageController,
+    private val messagesActionsUseCase: MessagesActionsUseCase,
     private val refreshController: RefreshController,
     private val sharedUtils: SharedUtils,
     private val snackbarManager: SnackbarManager,
@@ -175,15 +177,15 @@ class ThreadViewModel @Inject constructor(
     private val fakeReactions = MutableStateFlow<Map<String, Set<String>>>(emptyMap())
 
     /**
-    * Flow tracking the unsubscribe status of each message, keyed by its UID.
-    *
-    * Whenever a message is unsubscribed, its state (e.g., [UnsubscribeState.InProgress], [UnsubscribeState.Completed])
-    * is updated in this map, and the entire map is re-emitted to trigger a refresh of the [MessageUi].
-    *
-    * This flow is collected inside a [combine] to ensure that unsubscribe statuses are preserved
-    * even when other changes (e.g., sorting, filtering, content updates) trigger a full rebuild
-    * of the [MessageUi] list.
-    */
+     * Flow tracking the unsubscribe status of each message, keyed by its UID.
+     *
+     * Whenever a message is unsubscribed, its state (e.g., [UnsubscribeState.InProgress], [UnsubscribeState.Completed])
+     * is updated in this map, and the entire map is re-emitted to trigger a refresh of the [MessageUi].
+     *
+     * This flow is collected inside a [combine] to ensure that unsubscribe statuses are preserved
+     * even when other changes (e.g., sorting, filtering, content updates) trigger a full rebuild
+     * of the [MessageUi] list.
+     */
     private val unsubscribeStateByMessageUid = MutableStateFlow<Map<String, UnsubscribeState>>(emptyMap())
 
     @OptIn(ExperimentalCoroutinesApi::class)
@@ -416,7 +418,7 @@ class ThreadViewModel @Inject constructor(
     }
 
     private fun markThreadAsSeen(thread: Thread) = viewModelScope.launch(ioCoroutineContext) {
-        sharedUtils.markAsSeen(mailbox(), listOf(thread))
+        messagesActionsUseCase.markAsSeen(mailbox(), listOf(thread))
     }
 
     private fun sendMatomoAboutThreadMessagesCount(thread: Thread, featureFlags: Mailbox.FeatureFlagSet) {
